Move go plugins behind build tag
